1.0-alpha Release Refactor - Stable & Unstable branches (#28)
* Broke comment_parse into all of its grouped functions

* Removed comment_parse, centralized parameters being passed around everywhere

* Warning hunt

* Fix tip casing

* Fix tips recognized

* Fix walletsync

* {botname} where MoneroTipsBot is used in strings

* Made signature based on bot name, fixed walletsync

* Access denied fix on RPC

* Arguments fix

* Basic tip sending working again

* Bug fix with transaction amts

* ignore reddit timeouts

* wallet generation bugfixes

* walletsync/safewallet collision fix

* reddit timeout ignore

* prep for faucet merge

* Recursion fix & walletsync testnet fix

* Added timeout to broadcast tx function

* Made $ tips no longer display a billion decimal points

* Ignore another Reddit API execption

* Fixed get_info private keys

* Brought error messages back to public

* Don't load recipient RPC wallet if you don't need to

* More clear errors- tysm rbrunner :)

* remove debug

+++ tipbot/backend/safewallet.py
@@ -3,30 +3,49 @@ import os
 from monero.backends.jsonrpc import JSONRPCWallet
 from monero.wallet import Wallet
 
+import helper
 from logger import tipper_logger
-from wallet_rpc.rpc import RPC
+from tipbot.backend.rpc import RPC
 
 
-class safe_wallet(object):
+class SafeWallet(object):
 
     wallet = None
     rpc = None
+    password = None
+    timeout = None
 
-    def __init__(self, port, wallet_name, password, timeout=300):
+    def __init__(self, port, wallet_name, password=None, timeout=300):
         """
         Creates a monero-python Wallet based on the custom RPC that verifies it was created properly
 
         :param port: Port to open RPC
+        :param password: Password to open the wallet
         :param wallet_name: Lowercase string of username
         :param timeout: How long to let the RPC sync before killing
         """
+        if password is None:
+            password = helper.password
+
+        self.password = password
+        self.timeout = timeout
+
+        self.open_rpc(port=port, wallet_name=wallet_name, tries=5)
+
+
+
+    def open_rpc(self, port, wallet_name, password=password, timeout=timeout, tries=5):
+        if tries == 0:
+            tipper_logger.log(f"WARNING: FAILED to open {wallet_name}'s wallet!!")
+            return
         self.rpc = RPC(port=port, wallet_name=wallet_name, password=password, load_timeout=timeout)
 
         if not os.path.isfile("aborted-" + wallet_name):  # Check if wallet was emergency aborted
             self.wallet = Wallet(JSONRPCWallet(port=self.rpc.port, password=self.rpc.password, timeout=self.rpc.load_timeout))
         else:
-            tipper_logger.log("WARNING: " + wallet_name + " had their RPC aborted!!!")
+            tipper_logger.log(f"WARNING: {wallet_name} had their RPC aborted!!! Trying {tries} more times")
             os.remove("aborted-" + wallet_name)
+            self.open_rpc(port=port, wallet_name=wallet_name, password=password, timeout=timeout, tries=tries-1)
 
     def kill_rpc(self):
         self.rpc.kill()
